# acces to config.Global is protected with atomic.Value