--- a/api_definition.go
+++ b/api_definition.go
@@ -150,6 +150,7 @@ type APISpec struct {
 	ServiceRefreshInProgress bool
 	HTTPTransport            http.RoundTripper
 	HTTPTransportCreated     time.Time
+	GlobalConfig             config.Config
 }
 
 // APIDefinitionLoader will load an Api definition from a storage
@@ -176,8 +177,10 @@ func (a APIDefinitionLoader) MakeSpec(def *apidef.APIDefinition) *APISpec {
 	spec.SessionManager = &DefaultSessionManager{}
 	spec.OrgSessionManager = &DefaultSessionManager{}
 
+	spec.GlobalConfig = config.Global()
+
 	// Create and init the virtual Machine
-	if config.Global.EnableJSVM {
+	if config.Global().EnableJSVM {
 		spec.JSVM.Init(spec)
 	}
 
@@ -269,11 +272,11 @@ func (a APIDefinitionLoader) FromDashboardService(endpoint, secret string) []*AP
 	// Extract tagged entries only
 	apiDefs := make([]*apidef.APIDefinition, 0)
 
-	if config.Global.DBAppConfOptions.NodeIsSegmented {
-		tagList := make(map[string]bool, len(config.Global.DBAppConfOptions.Tags))
+	if config.Global().DBAppConfOptions.NodeIsSegmented {
+		tagList := make(map[string]bool, len(config.Global().DBAppConfOptions.Tags))
 		toLoad := make(map[string]*apidef.APIDefinition)
 
-		for _, mt := range config.Global.DBAppConfOptions.Tags {
+		for _, mt := range config.Global().DBAppConfOptions.Tags {
 			tagList[mt] = true
 		}
 
@@ -314,16 +317,16 @@ func (a APIDefinitionLoader) FromRPC(orgId string) []*APISpec {
 		return LoadDefinitionsFromRPCBackup()
 	}
 
-	store := RPCStorageHandler{UserKey: config.Global.SlaveOptions.APIKey, Address: config.Global.SlaveOptions.ConnectionString}
+	store := RPCStorageHandler{UserKey: config.Global().SlaveOptions.APIKey, Address: config.Global().SlaveOptions.ConnectionString}
 	if !store.Connect() {
 		return nil
 	}
 
 	// enable segments
 	var tags []string
-	if config.Global.DBAppConfOptions.NodeIsSegmented {
-		log.Info("Segmented node, loading: ", config.Global.DBAppConfOptions.Tags)
-		tags = config.Global.DBAppConfOptions.Tags
+	if config.Global().DBAppConfOptions.NodeIsSegmented {
+		log.Info("Segmented node, loading: ", config.Global().DBAppConfOptions.Tags)
+		tags = config.Global().DBAppConfOptions.Tags
 	}
 
 	apiCollection := store.GetApiDefinitions(orgId, tags)
@@ -349,7 +352,7 @@ func (a APIDefinitionLoader) processRPCDefinitions(apiCollection string) []*APIS
 	for _, def := range apiDefs {
 		def.DecodeFromDB()
 
-		if config.Global.SlaveOptions.BindToSlugsInsteadOfListenPaths {
+		if config.Global().SlaveOptions.BindToSlugsInsteadOfListenPaths {
 			newListenPath := "/" + def.Slug //+ "/"
 			log.Warning("Binding to ",
 				newListenPath,
@@ -684,7 +687,7 @@ func (a APIDefinitionLoader) compileURLRewritesPathSpec(paths []apidef.URLRewrit
 }
 
 func (a APIDefinitionLoader) compileVirtualPathspathSpec(paths []apidef.VirtualMeta, stat URLStatus, apiSpec *APISpec) []URLSpec {
-	if !config.Global.EnableJSVM {
+	if !config.Global().EnableJSVM {
 		return nil
 	}