--- a/coprocess.go
+++ b/coprocess.go
@@ -18,7 +18,7 @@ import (
 )
 
 var (
-	// EnableCoProcess will be overridden by config.Global.EnableCoProcess.
+	// EnableCoProcess will be overridden by config.Global().EnableCoProcess.
 	EnableCoProcess = false
 
 	// GlobalDispatcher will be implemented by the current CoProcess driver.
@@ -177,7 +177,7 @@ func (c *CoProcessor) ObjectPostProcess(object *coprocess.Object, r *http.Reques
 // CoProcessInit creates a new CoProcessDispatcher, it will be called when Tyk starts.
 func CoProcessInit() error {
 	var err error
-	if config.Global.CoProcessOptions.EnableCoProcess {
+	if config.Global().CoProcessOptions.EnableCoProcess {
 		GlobalDispatcher, err = NewCoProcessDispatcher()
 		EnableCoProcess = true
 	}
@@ -187,7 +187,7 @@ func CoProcessInit() error {
 // EnabledForSpec checks if this middleware should be enabled for a given API.
 func (m *CoProcessMiddleware) EnabledForSpec() bool {
 	// This flag is true when Tyk has been compiled with CP support and when the configuration enables it.
-	enableCoProcess := config.Global.CoProcessOptions.EnableCoProcess && EnableCoProcess
+	enableCoProcess := config.Global().CoProcessOptions.EnableCoProcess && EnableCoProcess
 	// This flag indicates if the current spec specifies any CP custom middleware.
 	var usesCoProcessMiddleware bool