refactoring for https://github.com/TykTechnologies/tyk/issues/1573

PR has lots of changes but the main idea is:
- global var `config.Global` and mutex `configMu` with logic around it are gone
- to read config - there is a new method `config.Global()` which returns `config.Config` and uses `sync/atomic`'s `Value` inside to provide atomic reads
- to write to config we should always have changed copy of it and the call `config.SetGlobal(conf Config)` which again provides atomic write
- all `APISpec` instances have a new field `GlobalConfig` which is set to `config.Global()` at API-load time and  is supposed to be used everywhere when we have `APISpec` instances - middle-wares and handlers to avoid to much calls to `config.Global()`

Also, maybe as a good side affect - we can introduce some alternative ways to do hot reload config with new version. 

I used our current benchmarking to see how we did:
```
➜  tyk git:(master) ./utils/ci-benchmark.sh > /tmp/old.txt
➜  tyk git:(master) git checkout config-global-races-fix
➜  tyk git:(config-global-races-fix) ./utils/ci-benchmark.sh > /tmp/new.txt
➜  tyk git:(config-global-races-fix) benchcmp /tmp/old.txt /tmp/new.txt    
benchmark                                               old ns/op     new ns/op     delta
BenchmarkApiReload-8                                    3403631       3742793       +9.96%
BenchmarkJWTSessionHMAC-8                               688676        714382        +3.73%
BenchmarkJWTSessionRSA-8                                859050        853650        -0.63%
BenchmarkJWTSessionRSABearer-8                          864115        845534        -2.15%
BenchmarkJWTSessionRSAWithRawSourceOnWithClientID-8     746854        732055        -1.98%
BenchmarkJWTSessionRSAWithRawSource-8                   737400        719759        -2.39%
BenchmarkJWTSessionRSAWithJWK-8                         745282        725662        -2.63%
BenchmarkJWTSessionRSAWithEncodedJWK-8                  739275        730120        -1.24%
BenchmarkCopyRequestResponse-8                          15344         15502         +1.03%

benchmark                                               old allocs     new allocs     delta
BenchmarkApiReload-8                                    23907          23907          +0.00%
BenchmarkJWTSessionHMAC-8                               714            713            -0.14%
BenchmarkJWTSessionRSA-8                                792            791            -0.13%
BenchmarkJWTSessionRSABearer-8                          793            793            +0.00%
BenchmarkJWTSessionRSAWithRawSourceOnWithClientID-8     814            814            +0.00%
BenchmarkJWTSessionRSAWithRawSource-8                   797            797            +0.00%
BenchmarkJWTSessionRSAWithJWK-8                         801            801            +0.00%
BenchmarkJWTSessionRSAWithEncodedJWK-8                  805            805            +0.00%
BenchmarkCopyRequestResponse-8                          76             76             +0.00%

benchmark                                               old bytes     new bytes     delta
BenchmarkApiReload-8                                    1789646       1789704       +0.00%
BenchmarkJWTSessionHMAC-8                               117808        117699        -0.09%
BenchmarkJWTSessionRSA-8                                161132        160781        -0.22%
BenchmarkJWTSessionRSABearer-8                          162085        161953        -0.08%
BenchmarkJWTSessionRSAWithRawSourceOnWithClientID-8     140062        140033        -0.02%
BenchmarkJWTSessionRSAWithRawSource-8                   132024        131912        -0.08%
BenchmarkJWTSessionRSAWithJWK-8                         132034        131847        -0.14%
BenchmarkJWTSessionRSAWithEncodedJWK-8                  132030        131888        -0.11%
BenchmarkCopyRequestResponse-8                          35584         35584         +0.00%
``` 

As we see:
- `B/op` and `allocs/op` are OK or maybe even better
-  `ns/op` - `BenchmarkApiReload` increased  `+9.96%` which I think is expected, however, I think we can do better
-  `ns/op` - JWT middleware looks OK except `BenchmarkJWTSessionHMAC`'s `+3.73%` - needs to be investigated