@matiasinsaurralde 2018-04-05
LGTM, haven't checked the issue with the tests yet