@matiasinsaurralde 2018-05-04
Tests look ok now however there are some (new) code conflicts