@buger 2018-05-04
@dencoded lets fix conflict and merge it asap. Sorry for delay.